Fix intermittent failure in executedEvery2Sec #10682
